delete PublicCardList and use inline
